chore: separate columns definition from table component
